% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstd.R
\name{zstdDecompress}
\alias{zstdDecompress}
\title{Decopresses binary data}
\usage{
zstdDecompress(data)
}
\arguments{
\item{data}{compressed data to be decompressed}
}
\value{
decompressed data
}
\description{
Decopresses binary data
}
