% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstd.R
\name{zstdMaxCLevel}
\alias{zstdMaxCLevel}
\title{Returns maximum compression level supported by zstandard}
\usage{
zstdMaxCLevel()
}
\value{
maximum supported compression level
}
\description{
Returns maximum compression level supported by zstandard
}
\seealso{
\link{memCompress} \link{zstdCompress} \link{zstdDecompress}
}
