% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstdR.R
\docType{package}
\name{zstdR}
\alias{zstdR}
\alias{compressZstd}
\alias{compressZstd.default}
\alias{compressZstd.raw}
\alias{uncompressZstd}
\alias{uncompressZstd.raw}
\title{Zstandard, or zstd, compression algorithm}
\usage{
compressZstd(object, level, ...)

\method{compressZstd}{default}(object, level = 3, ...)

\method{compressZstd}{raw}(object, level = 3, ...)

uncompressZstd(object, ...)

\method{uncompressZstd}{raw}(object, ...)
}
\arguments{
\item{object}{Object to (un)compress.
It will serialize the object on compression if it's not a raw vector.}

\item{level}{Integer. Compression level.}
}
\value{
\code{compress} returns the raw snappy encoded bytes of
     the text input.

  \code{compress_raw} returns the raw snappy encoded bytes
     of the raw input.

  \code{decompress} returns the a character vector made
     from the decoded bytes of its input. If this coercion
     can't be performed (i.e., the bytes can't be represented
     as a string) an error will be thrown.

  \code{decompress_raw} returns the a raw bytes decoded from
     its input. When passed bytes which are not snappy
     compressed, this function will return NULL.
}
\description{
Zstandard, or zstd as short version, is a fast lossless compression algorithm
  created by Facebook Inc.

  This package provides bindings to the Zstandard C implementation.
}
\examples{
library(zstdR)
# Roundtrip compression of character vector
decompress( compress( "this and that" ) ) # "this and that"

compress("")                              # 00
compress_raw(c())                         # 00

decompress_raw(0)                         # raw(0)
decompress(0)                             # ""

}
