% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstd.R
\name{zstdCompress}
\alias{zstdCompress}
\title{Compresses character or binary vector}
\usage{
zstdCompress(data, level = 3)
}
\arguments{
\item{data}{input data to be compressed}

\item{level}{compression level to use, default is 3}
}
\value{
compressed data
}
\description{
Compresses character or binary vector
}
